APT-29-CozyBear/requirements.txt
pyautogui
pyvirtualdisplay
requests
pycryptodome

